# perf(tspath): optimize the fast path case for `ToFileNameLowerCase`