When using project services, `ToFileNameLowerCase` is called 10, 312, 056 times on various paths, this means that it's incredibly hot path.

Since most projects won't hav e non-ascii filenames, we can optimize for this case. This yields up to a 5x perf improvement in some scenarios.

Please feel free to close, if the perf degregation againt non-ascii file paths outweighs the perf gains for the common case

| Test Case                     | Likelihood Category | Old (ns) | New (ns) |
| ----------------------------- | ------------------- | -------- | -------- |
| All lowercase ASCII short     | High                | 45.1     | 8.5      |
| All uppercase ASCII short     | Medium              | 71.8     | 37.9     |
| Mixed case ASCII short        | High                | 62.2     | 41.9     |
| Mixed case ASCII longer       | High                | 134.0    | 60.9     |
| Non-ASCII without İ (e.g., ß) | Low                 | 152.0    | 166.7    |
| Non-ASCII with İ              | Very low            | 147.5    | 156.8    |
| Non-ASCII with ı              | Low                 | 128.2    | 138.1    |
| Long mixed ASCII              | Low                 | 1260.6   | 547.8    |