@jakebailey 2025-08-11
Ah, yes. I didn't realize that's what you were doing. I think I was expecting results through https://pkg.go.dev/golang.org/x/perf/benchstat.

Comparing main and this PR with `go test -run=- -bench='BenchmarkToFileNameLowerCase' -benchmem -count=10 ./internal/tspath`:

```
goos: linux
goarch: amd64
pkg: github.com/microsoft/typescript-go/internal/tspath
cpu: Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz
                                                     │   old.txt    │               new.txt               │
                                                     │    sec/op    │   sec/op     vs base                │
ToFileNameLowerCase//path/to/file.ext-20               58.495n ± 2%   9.892n ± 1%  -83.09% (p=0.000 n=10)
ToFileNameLowerCase//PATH/TO/FILE.EXT-20               101.25n ± 1%   45.84n ± 0%  -54.73% (p=0.000 n=10)
ToFileNameLowerCase//path/to/FILE.EXT-20                85.81n ± 2%   46.68n ± 1%  -45.60% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc-20      186.45n ± 1%   97.24n ± 1%  -47.84% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc#01-20    206.2n ± 1%   222.8n ± 1%   +8.05% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc#02-20    202.5n ± 1%   215.2n ± 1%   +6.25% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc#03-20    174.4n ± 1%   190.2n ± 1%   +9.06% (p=0.000 n=10)
ToFileNameLowerCase/FoO/FoO/FoO/FoO/FoO/...etc-20      1623.5n ± 1%   952.5n ± 1%  -41.33% (p=0.000 n=10)
geomean                                                 180.4n        107.5n       -40.38%
```
```
                                                     │   old.txt    │                new.txt                │
                                                     │     B/op     │    B/op     vs base                   │
ToFileNameLowerCase//path/to/file.ext-20               0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//PATH/TO/FILE.EXT-20               24.00 ± 0%     24.00 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//path/to/FILE.EXT-20               24.00 ± 0%     24.00 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc-20      48.00 ± 0%     96.00 ± 0%  +100.00% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc#01-20   48.00 ± 0%     48.00 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc#02-20   48.00 ± 0%     48.00 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc#03-20   48.00 ± 0%     48.00 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase/FoO/FoO/FoO/FoO/FoO/...etc-20      416.0 ± 0%     832.0 ± 0%  +100.00% (p=0.000 n=10)
geomean                                                           ²                +18.92%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```
```
                                                     │   old.txt    │                new.txt                │
                                                     │  allocs/op   │ allocs/op   vs base                   │
ToFileNameLowerCase//path/to/file.ext-20               0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//PATH/TO/FILE.EXT-20               1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//path/to/FILE.EXT-20               1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc-20      1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
ToFileNameLowerCase//user/UserName/proje...etc#01-20   1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc#02-20   1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase//user/UserName/proje...etc#03-20   1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToFileNameLowerCase/FoO/FoO/FoO/FoO/FoO/...etc-20      1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
geomean                                                           ²                +18.92%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Which I think matches what you were quoting. Surprised the other cases get slower, honestly.