@jakebailey 2025-08-11
> yess, although I can write go, it's been a while and I am far from an expert, i'm also not up to date with the latest tooling, so there's definitly cases where I could be using a tool to do what i'm doing manually 😅

For the record, I do:

```console
$ go test -run=- -bench='BenchmarkToFileNameLowerCase' -benchmem -count=10 ./internal/tspath | tee old.txt
# switch branch
$ go test -run=- -bench='BenchmarkToFileNameLowerCase' -benchmem -count=10 ./internal/tspath | tee new.txt
$ benchstat old.txt new.txt
```