@jakebailey 2025-08-12
Just out of curiosity, where did you experience this being a bottleneck?