@Boshen 2025-08-12
> Just out of curiosity, where did you experience this being a bottleneck?

We are diving into

* https://github.com/oxc-project/tsgolint/issues/67 

i.e. running tsgolint in large monorepos (1000+ files).