@camc314 2025-08-12
> Just out of curiosity, where did you experience this being a bottleneck?

Yep to add to Boshen's point, i'm testing this on vscode repo (~6100) files, during which, this function is called 10 million + time.

thanks for the extra bechmarking info Jake, i'll be using that next time!